# MyApi/Models/Course.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;

namespace MyApi.Models
{
    public class Course
    {
        public string course_name{ get; set; } = "";

    }
}