# MyApi/Models/Departments.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Web;

namespace MyApi.Models
{
    public class Departments
    {
        public string department_id { get; set; } = "";
        public string department_name { get; set; } = "";
        
    }
}